+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+  - Fixed `LasZipDecompressor` to still be able to read points even when the source
+    is not seekable. Seeking is only required to get the ChunkTable, and the ChunkTable
+    is only required when wanting to use `LasZipDecompressor.seek` to get a point.
+
 # 0.6.1
   - Added support for `PointWise` compressed data in `LasZipDecompressor`.
   - Fixed `LasZipCompressor` when no points where compressed.